added scraping (#2)
Quote by Khaled Kanawati : "As discussed, I scraped the Lebanon debate website and accordingly generated a list of incidents (dicts), each of which contains: title, url, content, and id, with id being a uuid. The reason I am even providing a uuid is for db purposes as from previosu experience DB needs an id. URL will be useful as to provide a src and prevent duplicates on db. For now I am not saving the result (source_incidents in get_feed) anywhere as I do not have much knowledge on how to use Dynamo, if u want just let me know if u want me to do that as to save it to db, where all u have to do is create a new db section for articles.
One issue I found though is that now we have to differentiate between crime and non-crime related stuff, I am currently experimenting on that using Gemini and will let you know what I reach. Therefore I would scrape, determine if crime, if yes, save to db, if no, disregard."